Refactor folders and update README.md
