+++ Data Strcutures/String Occurence.cpp
