+++ Data Strcutures/magic_square.c
