+++ Python/wordle/wordfile.json
