--- languse
@@ -1,38 +1,0 @@
-<div class="language-selector">
-    <label for="lang">Language:</label>
-    <select id="lang" onchange="setLanguage(this.value)">
-        <option value="en">English</option>
-        <option value="es">Spanish</option>
-        <option value="fr">French</option>
-        <option value="zh">Mandarin</option>
-    </select>
-</div>
-
-
-const translations = {
-    en: {
-        title: "Welcome to HacktoberWall",
-        description: "A platform celebrating open-source contributions.",
-    },
-    es: {
-        title: "Bienvenido a HacktoberWall",
-        description: "Una plataforma que celebra las contribuciones de código abierto.",
-    },
-    // Add other languages...
-};
-
-function setLanguage(lang) {
-    document.getElementById("site-title").innerText = translations[lang].title;
-    document.getElementById("site-description").innerText = translations[lang].description;
-}
-
-
-
-function setLanguage(lang) {
-    localStorage.setItem("preferredLanguage", lang);
-    // Update content as shown above...
-}
-
-// On page load, set the language based on user preference
-const preferredLang = localStorage.getItem("preferredLanguage") || "en";
-setLanguage(preferredLang);
